feat(web): unify workspace directory picker across settings and chat
Extract shared fs picker components for chat, cron, and agent defaults, add Electron defaultPath support, and a shortcut to ~/.xopc/workspace.
